% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_mm}
\alias{combine_mm}
\title{Combine multimedia files in a folder}
\usage{
combine_mm(folder, multimedia = ".mp3")
}
\arguments{
\item{folder}{character}

\item{multimedia}{character, specify the multimedia type, i.e. ".mp3", ".mp4", ".wav".}
}
\value{
character
}
\description{
Combine multimedia files in a folder
}
\examples{
# folder named "Heart_of_Darkness"
x = combine_mm("Heart_of_Darkness")
x # able to see the Powershell output
}
